requirements.txt
flask
flask-cors
groq
google-generativeai
python-dotenv
pandas
numpy
faiss-cpu
sentence-transformers
markdown
